requirements.txt
streamlit==1.40.0
openai==1.58.1
python-dotenv==1.0.1